+++ noodles-bam/src/reader.rs
@@ -15,7 +15,7 @@ use std::{
 
 use byteorder::{LittleEndian, ReadBytesExt};
 use noodles_bgzf as bgzf;
-use noodles_core::{region::Interval, Position, Region};
+use noodles_core::{Position, Region};
 use noodles_csi::{binning_index::ReferenceSequenceExt, BinningIndex};
 use noodles_fasta as fasta;
 use noodles_sam::{
@@ -301,7 +301,7 @@ where
     ///
     /// let reference_sequences = header.reference_sequences();
     /// let index = bai::read("sample.bam.bai")?;
-    /// let region = Region::new("sq0", 17711..=28657);
+    /// let region = "sq0:8-13".parse()?;
     /// let query = reader.query(reference_sequences, &index, &region)?;
     ///
     /// for result in query {
@@ -310,24 +310,20 @@ where
     /// }
     /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
+    #[allow(clippy::type_complexity)]
     pub fn query<I, RS>(
         &mut self,
         reference_sequences: &ReferenceSequences,
         index: &I,
         region: &Region,
-    ) -> io::Result<Query<'_, R, Interval>>
+    ) -> io::Result<Query<'_, R, (Bound<i32>, Bound<i32>)>>
     where
         I: BinningIndex<RS>,
         RS: ReferenceSequenceExt,
     {
         let (reference_sequence_id, interval) = resolve_region(reference_sequences, region)?;
-        let chunks = index.query(reference_sequence_id, interval)?;
-        Ok(Query::new(
-            self,
-            chunks,
-            reference_sequence_id,
-            region.interval(),
-        ))
+        let chunks = index.query(reference_sequence_id, region.interval())?;
+        Ok(Query::new(self, chunks, reference_sequence_id, interval))
     }
 
     /// Returns an iterator of unmapped records after querying for the unmapped region.
@@ -490,11 +486,11 @@ pub(crate) fn bytes_with_nul_to_string(buf: &[u8]) -> io::Result<String> {
 pub(crate) fn resolve_region(
     reference_sequences: &ReferenceSequences,
     region: &Region,
-) -> io::Result<(usize, (Bound<Position>, Bound<Position>))> {
-    fn cast_bound_i32_to_bound_position(bound: Bound<i32>) -> Bound<Position> {
+) -> io::Result<(usize, (Bound<i32>, Bound<i32>))> {
+    fn cast_bound_position_to_bound_i32(bound: Bound<Position>) -> Bound<i32> {
         match bound {
-            Bound::Included(n) => Bound::Included(Position::try_from(n as usize).unwrap()),
-            Bound::Excluded(n) => Bound::Excluded(Position::try_from(n as usize).unwrap()),
+            Bound::Included(position) => Bound::Included(usize::from(position) as i32),
+            Bound::Excluded(position) => Bound::Excluded(usize::from(position) as i32),
             Bound::Unbounded => Bound::Unbounded,
         }
     }
@@ -512,8 +508,8 @@ pub(crate) fn resolve_region(
         })?;
 
     let interval = (
-        cast_bound_i32_to_bound_position(region.start()),
-        cast_bound_i32_to_bound_position(region.end()),
+        cast_bound_position_to_bound_i32(region.start()),
+        cast_bound_position_to_bound_i32(region.end()),
     );
 
     Ok((i, interval))